Issue #1031. Move error handling to its own blueprint module.
