+++ webcompat/__init__.py
@@ -28,9 +28,10 @@ import webcompat.views
 # register blueprints
 from api.endpoints import api
 from api.uploads import uploads
+from error_handlers import error_handlers
 from webhooks import webhooks
 
-for blueprint in [api, uploads, webhooks]:
+for blueprint in [api, error_handlers, uploads, webhooks]:
     app.register_blueprint(blueprint)
